Do not show the "log back in" button if the url is blank
